+++ packages/amplify-ui-components/src/components/amplify-require-new-password/amplify-require-new-password.tsx
@@ -108,7 +108,10 @@ export class AmplifyRequireNewPassword {
         handleSubmit={this.handleSubmit}
         loading={this.loading}
         secondaryFooterContent={
-          <amplify-button variant="anchor" onClick={() => this.handleAuthStateChange(AuthState.SignIn)}>
+          <amplify-button
+            variant="anchor"
+            onClick={() => this.handleAuthStateChange(AuthState.SignIn)}
+          >
             {I18n.get(Translations.BACK_TO_SIGN_IN)}
           </amplify-button>
         }