+++ packages/aws-amplify-react/src/Auth/index.tsx
@@ -29,6 +29,7 @@ export { TOTPSetup } from './TOTPSetup';
 export { Loading } from './Loading';
 
 export * from './Provider';
+export * from './common/types';
 
 export function withAuthenticator(
 	Comp,