+++ packages/datastore/__tests__/DataStore.ts
@@ -1,29 +1,47 @@
 import 'fake-indexeddb/auto';
-import * as uuidValidate from 'uuid-validate';
-import { initSchema as initSchemaType } from '../src/datastore/datastore';
+import uuidValidate from 'uuid-validate';
+import {
+	initSchema as initSchemaType,
+	DataStore as DataStoreType,
+} from '../src/datastore/datastore';
 import {
 	ModelInit,
 	MutableModel,
 	PersistentModelConstructor,
 	Schema,
+	NonModelTypeConstructor,
 } from '../src/types';
+import { ExclusiveStorage as StorageType } from '../src/storage/storage';
+import Observable from 'zen-observable-ts';
 
 let initSchema: typeof initSchemaType;
+let DataStore: typeof DataStoreType;
 
 beforeEach(() => {
 	jest.resetModules();
 
-	({ initSchema } = require('../src/datastore/datastore'));
+	jest.doMock('../src/storage/storage', () => {
+		const mock = jest.fn().mockImplementation(() => ({
+			runExclusive: jest.fn(),
+			query: jest.fn(),
+			observe: jest.fn(() => Observable.of()),
+		}));
+
+		(<any>mock).getNamespace = () => ({ models: {} });
+
+		return { ExclusiveStorage: mock };
+	});
+	({ initSchema, DataStore } = require('../src/datastore/datastore'));
 });
 
 describe('DataStore tests', () => {
 	describe('initSchema tests', () => {
-		test('Class is created', () => {
+		test('Model class is created', () => {
 			const classes = initSchema(testSchema());
 
 			expect(classes).toHaveProperty('Model');
 
-			const { Model } = classes;
+			const { Model } = classes as { Model: PersistentModelConstructor<Model> };
 
 			let property: keyof PersistentModelConstructor<any> = 'copyOf';
 			expect(Model).toHaveProperty(property);
@@ -31,7 +49,7 @@ describe('DataStore tests', () => {
 			expect(typeof Model.copyOf).toBe('function');
 		});
 
-		test('Class can be instantiated', () => {
+		test('Model class can be instantiated', () => {
 			const { Model } = initSchema(testSchema()) as {
 				Model: PersistentModelConstructor<Model>;
 			};
@@ -65,6 +83,40 @@ describe('DataStore tests', () => {
 			expect(
 				uuidValidate(model.id.replace(/^(.{4})-(.{4})-(.{8})/, '$3-$2-$1'), 1)
 			).toBe(true);
+		});
+
+		test('initSchema is executed only once', () => {
+			initSchema(testSchema());
+
+			expect(() => {
+				initSchema(testSchema());
+			}).toThrow('The schema has already been initialized');
+		});
+
+		test('Non @model class is created', () => {
+			const classes = initSchema(testSchema());
+
+			expect(classes).toHaveProperty('Metadata');
+
+			const { Metadata } = classes;
+
+			let property: keyof PersistentModelConstructor<any> = 'copyOf';
+			expect(Metadata).not.toHaveProperty(property);
+		});
+
+		test('Non @model class can be instantiated', () => {
+			const { Metadata } = initSchema(testSchema()) as {
+				Metadata: NonModelTypeConstructor<Metadata>;
+			};
+
+			const metadata = new Metadata({
+				author: 'some author',
+				tags: [],
+			});
+
+			expect(metadata).toBeInstanceOf(Metadata);
+
+			expect(metadata).not.toHaveProperty('id');
 		});
 	});
 
@@ -121,6 +173,100 @@ describe('DataStore tests', () => {
 			// ID should be kept the same
 			expect(model1.id).toBe(model2.id);
 		});
+
+		test('Non @model - Field cannot be changed', () => {
+			const { Metadata } = initSchema(testSchema()) as {
+				Metadata: NonModelTypeConstructor<Metadata>;
+			};
+
+			const nonModel = new Metadata({
+				author: 'something',
+			});
+
+			expect(() => {
+				(<any>nonModel).author = 'edit';
+			}).toThrowError("Cannot assign to read only property 'author' of object");
+		});
+	});
+
+	describe('Initialization', () => {
+		test('start is called only once', async () => {
+			const storage: StorageType = require('../src/storage/storage')
+				.ExclusiveStorage;
+
+			const classes = initSchema(testSchema());
+
+			const { Model } = classes as { Model: PersistentModelConstructor<Model> };
+
+			const promises = [
+				DataStore.query(Model),
+				DataStore.query(Model),
+				DataStore.query(Model),
+				DataStore.query(Model),
+			];
+
+			await Promise.all(promises);
+
+			expect(storage).toHaveBeenCalledTimes(1);
+		});
+
+		test('It is initialized when observing (no query)', async () => {
+			const storage: StorageType = require('../src/storage/storage')
+				.ExclusiveStorage;
+
+			const classes = initSchema(testSchema());
+
+			const { Model } = classes as { Model: PersistentModelConstructor<Model> };
+
+			DataStore.observe(Model).subscribe(jest.fn());
+
+			expect(storage).toHaveBeenCalledTimes(1);
+		});
+	});
+
+	describe('Basic operations', () => {
+		test('Save returns the saved model', async () => {
+			let model: Model;
+
+			jest.resetModules();
+			jest.doMock('../src/storage/storage', () => {
+				const mock = jest.fn().mockImplementation(() => ({
+					runExclusive: jest.fn(() => [model]),
+				}));
+
+				(<any>mock).getNamespace = () => ({ models: {} });
+
+				return { ExclusiveStorage: mock };
+			});
+			({ initSchema, DataStore } = require('../src/datastore/datastore'));
+
+			const classes = initSchema(testSchema());
+
+			const { Model } = classes as { Model: PersistentModelConstructor<Model> };
+
+			model = new Model({
+				field1: 'Some value',
+			});
+
+			const result = await DataStore.save(model);
+
+			expect(result).toMatchObject(model);
+		});
+	});
+
+	test("non-@models can't be saved", async () => {
+		const { Metadata } = initSchema(testSchema()) as {
+			Metadata: NonModelTypeConstructor<Metadata>;
+		};
+
+		const metadata = new Metadata({
+			author: 'some author',
+			tags: [],
+		});
+
+		await expect(DataStore.save(<any>metadata)).rejects.toThrow(
+			'Object is not an instance of a valid model'
+		);
 	});
 });
 
@@ -129,6 +275,7 @@ describe('DataStore tests', () => {
 declare class Model {
 	public readonly id: string;
 	public readonly field1: string;
+	public readonly metadata?: Metadata;
 
 	constructor(init: ModelInit<Model>);
 
@@ -138,12 +285,19 @@ declare class Model {
 	): Model;
 }
 
+export declare class Metadata {
+	readonly author: string;
+	readonly tags?: string[];
+	constructor(init: Metadata);
+}
+
 function testSchema(): Schema {
 	return {
 		enums: {},
 		models: {
 			Model: {
 				name: 'Model',
+				pluralName: 'Models',
 				syncable: true,
 				fields: {
 					id: {
@@ -158,10 +312,20 @@ function testSchema(): Schema {
 						type: 'String',
 						isRequired: true,
 					},
+					metadata: {
+						name: 'metadata',
+						isArray: false,
+						type: {
+							nonModel: 'Metadata',
+						},
+						isRequired: false,
+						attributes: [],
+					},
 				},
 			},
 			LocalModel: {
 				name: 'LocalModel',
+				pluralName: 'LocalModels',
 				syncable: false,
 				fields: {
 					id: {
@@ -179,7 +343,28 @@ function testSchema(): Schema {
 				},
 			},
 		},
-		version: 1,
+		nonModels: {
+			Metadata: {
+				name: 'Metadata',
+				fields: {
+					author: {
+						name: 'author',
+						isArray: false,
+						type: 'String',
+						isRequired: true,
+						attributes: [],
+					},
+					tags: {
+						name: 'tags',
+						isArray: true,
+						type: 'String',
+						isRequired: false,
+						attributes: [],
+					},
+				},
+			},
+		},
+		version: '1',
 	};
 }