+++ packages/analytics/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@2.2.6...@aws-amplify/analytics@2.2.7) (2020-03-25)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+## [2.2.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@2.2.5...@aws-amplify/analytics@2.2.6) (2020-02-28)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+## [2.2.5](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@2.2.4...@aws-amplify/analytics@2.2.5) (2020-02-14)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
 ## [2.2.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@2.2.2...@aws-amplify/analytics@2.2.4) (2020-02-07)
 
 ### Bug Fixes