+++ packages/datastore/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@1.0.8...@aws-amplify/datastore@1.1.0) (2020-03-25)
+
+### Bug Fixes
+
+- **@aws-amplify/datastore:** Fix query and delete types ([#5032](https://github.com/aws-amplify/amplify-js/issues/5032)) ([fdca554](https://github.com/aws-amplify/amplify-js/commit/fdca5541372662ffa1d932b665c481a78e4ccdc7)), closes [#4827](https://github.com/aws-amplify/amplify-js/issues/4827)
+- **@aws-amplify/datastore:** Storage should be re-initialized after DataStore.clear() ([#5083](https://github.com/aws-amplify/amplify-js/issues/5083)) ([0ddb6af](https://github.com/aws-amplify/amplify-js/commit/0ddb6af3163fc624cc4f320ecf2b2463d7d6b102)), closes [#5076](https://github.com/aws-amplify/amplify-js/issues/5076)
+
+### Features
+
+- **@aws-amplify/datastore:** Support non-[@model](https://github.com/model) types in DataStore ([#5128](https://github.com/aws-amplify/amplify-js/issues/5128)) ([b884ea2](https://github.com/aws-amplify/amplify-js/commit/b884ea2ce730d8ce981a5921f74f8f37338f6f42))
+
+## [1.0.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@1.0.7...@aws-amplify/datastore@1.0.8) (2020-02-28)
+
+**Note:** Version bump only for package @aws-amplify/datastore
+
+## [1.0.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@1.0.6...@aws-amplify/datastore@1.0.7) (2020-02-14)
+
+**Note:** Version bump only for package @aws-amplify/datastore
+
 ## [1.0.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore@1.0.4...@aws-amplify/datastore@1.0.6) (2020-02-07)
 
 **Note:** Version bump only for package @aws-amplify/datastore